Refactor test_import_unmapping.py
