--- test/integration/ggrc/test_csvs/multi_basic_policy_orggroup_product_with_mappings.csv
@@ -1,26 +1,0 @@
-Object type,,,,,,
-product,code,title,admin,verifier,assignee,
-,pro 1,meatloaf,user@example.com,user@example.com,user@example.com,
-,pro 2,pastrami,user@example.com,user@example.com,user@example.com,
-,pro 3,landjaeger,user@example.com,user@example.com,user@example.com,
-,pro 4,ham,user@example.com,user@example.com,user@example.com,
-,pro 5,bacon,user@example.com,user@example.com,user@example.com,
-,,,,,,
-Object type,,,,,,
-Org group,code,title,map:product,admin,verifier,assignee
-,org-1,boudin,pro 4,user@example.com,user@example.com,user@example.com
-,org-2,tail,pro 5,user@example.com,user@example.com,user@example.com
-,org-3,filet,,user@example.com,user@example.com,user@example.com
-,org-4,mignon,"pro 1
-pro 4
-pro 3",user@example.com,user@example.com,user@example.com
-,,,,,,
-Object type,,,,,,
-Policy,code,title,Map: Org group,map:product,admin,
-,p-1,Bacon,"org-1
-org-4",pro 4,user@example.com,
-,p-2,ipsum,"org-1
-org-4",,user@example.com,
-,p-3,dolor,"org-1
-org-4",,user@example.com,
-,p-4,amet,org-1,,user@example.com,
