--- test/integration/ggrc/test_csvs/multi_basic_policy_orggroup_product_with_unmappings.csv
@@ -1,29 +1,0 @@
-"Object type",,,,,
-Policy,code,title,"unmap: Org group",,admin
-,p-1,Bacon,,,user@example.com
-,p-2,ipsum,org-1,,user@example.com
-,p-3,dolor,org-1,,user@example.com
-,p-4,amet,org-1,,user@example.com
-,,,,,
-,,,,,
-,,,,,
-"Object type",,,,,
-"Org group",code,title,unmap:Policy,Unmap:product,admin
-,org-1,boudin,"p-1
-p-2
-p-3","pro 4",user@example.com
-,org-2,tail,,"pro 5",user@example.com
-,org-3,filet,,,user@example.com
-,org-4,mignon,"p-1
-p-2
-p-3","pro 1
-pro 4
-pro 3",user@example.com
-,,,,,
-"Object type",,,,,
-product,code,title,"unmap: org group",unmap:policy,admin
-,"pro 1",meatloaf,,,user@example.com
-,"pro 2",pastrami,,,user@example.com
-,"pro 3",landjaeger,,,user@example.com
-,"pro 4",ham,org-4,p-1,user@example.com
-,"pro 5",bacon,,,user@example.com
